requirements.txt
fastapi==0.115.12
gunicorn==23.0.0
psycopg[binary]==3.1.12
pydantic==2.11.3
python-decouple==3.8
requests==2.32.3
SQLAlchemy==2.0.40
sqlmodel==0.0.24
uvicorn==0.34.1
